Move Dockerfile to top level of repository.
In this commit the Dockerfile for our project is moved to the top level
of the repository. The content of the file is changed to reflect the
new relative paths.
This is changed so that our project can easily be executed using Docker
without having to move around our directory structure.