Add template code to src
With this commit we add the cpp-opencv template for working
with rec-files to the src/ directory of our project.
This commit should not be applied alone as it does not include
the necessary libraries. However, since neither the Dockerfile, nor
CMakeLists.txt are changed at this point, applying it is save.

Refer to:
https://github.com/chalmers-revere/working-with-rec-files/tree/master/cpp-opencv
for more on the template.